+++ test/dummy/config/routes.rb
@@ -1,5 +1,6 @@
 Rails.application.routes.draw do
   root 'application#cowsay'
   get '/moo', to: 'application#cow'
+  get '/normal', to: 'application#normal'
   mount RenderCow::Engine => '/render_cow'
 end