+++ test/integration/navigation_test.rb
@@ -5,15 +5,20 @@ class NavigationTest < ActionDispatch::IntegrationTest
     RenderCow.moo('Moo')
   end
 
-  test 'access cow path' do
-    get '/'
-    assert_response :success
-    assert_equal response.body, default_moo
-  end
+  # test 'access cow path' do
+  #   get '/'
+  #   assert_response :success
+  #   assert_equal response.body, default_moo
+  # end
 
-  test 'can also just use cow:' do
-    get '/moo'
-    assert_response :success
-    assert_equal response.body, default_moo
+  # test 'can also just use cow:' do
+  #   get '/moo'
+  #   assert_response :success
+  #   assert_equal response.body, default_moo
+  # end
+
+  test 'can render normal views' do
+    get '/normal'
+    assert_includes response.body, 'Hello'
   end
 end