+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0
+
+A new and exciting feature! 
+You can now render your existing views! 
+Previous versions would only let you render cowsays. 
+So a bugfix. 
 ## 0.1.1
 
 * Relax version constraint as my Rails 7 project was not able to use RenderCow.